Add Review suggestions
# Conflicts:
#	packages/page/src/Infrastructure/Symfony/HttpKernel/SuluPageBundle.php
