+++ packages/page/src/Application/MessageHandler/RemovePageMessageHandler.php
@@ -12,22 +12,15 @@
 namespace Sulu\Page\Application\MessageHandler;
 
 use Sulu\Bundle\ActivityBundle\Application\Collector\DomainEventCollectorInterface;
-use Sulu\Bundle\SecurityBundle\System\SystemStoreInterface;
 use Sulu\Bundle\TrashBundle\Application\TrashManager\TrashManagerInterface;
-use Sulu\Component\Rest\Exception\InsufficientDescendantPermissionsException;
-use Sulu\Component\Security\Authentication\UserInterface;
-use Sulu\Component\Security\Authorization\AccessControl\AccessControlRepositoryInterface;
-use Sulu\Component\Security\Authorization\PermissionTypes;
 use Sulu\Content\Domain\Model\DimensionContentCollection;
 use Sulu\Page\Application\Message\RemovePageMessage;
 use Sulu\Page\Domain\Event\PageRemovedEvent;
 use Sulu\Page\Domain\Exception\RemovePageDependantResourcesFoundException;
-use Sulu\Page\Domain\Model\Page;
 use Sulu\Page\Domain\Model\PageDimensionContent;
 use Sulu\Page\Domain\Model\PageDimensionContentInterface;
 use Sulu\Page\Domain\Model\PageInterface;
 use Sulu\Page\Domain\Repository\PageRepositoryInterface;
-use Symfony\Bundle\SecurityBundle\Security;
 
 /**
  * @experimental
@@ -37,16 +30,9 @@ use Symfony\Bundle\SecurityBundle\Security;
  */
 final class RemovePageMessageHandler
 {
-    /**
-     * @param array<string, int> $permissions
-     */
     public function __construct(
         private PageRepositoryInterface $pageRepository,
         private DomainEventCollectorInterface $domainEventCollector,
-        private AccessControlRepositoryInterface $accessControlRepository,
-        private SystemStoreInterface $systemStore,
-        private ?Security $security,
-        private array $permissions,
         private ?TrashManagerInterface $trashManager = null,
     ) {
     }
@@ -54,8 +40,6 @@ final class RemovePageMessageHandler
     public function __invoke(RemovePageMessage $message): void
     {
         $page = $this->pageRepository->getOneBy($message->getIdentifier());
-
-        $this->checkPermissionsForDescendants($page);
 
         if (!$message->getForceRemoveChildren()) {
             $this->checkDescendantsForDependencyWarning($page);
@@ -92,36 +76,6 @@ final class RemovePageMessageHandler
             $title,
             $context
         ));
-    }
-
-    private function checkPermissionsForDescendants(PageInterface $page): void
-    {
-        $user = $this->getCurrentUser();
-        if (null === $user) {
-            return;
-        }
-
-        $descendantIds = $this->pageRepository->findDescendantIdsById($page->getUuid());
-        if (empty($descendantIds)) {
-            return;
-        }
-
-        $authorizedDescendantIds = $this->accessControlRepository->findIdsWithGrantedPermissions(
-            $user,
-            $this->permissions[PermissionTypes::DELETE],
-            Page::class,
-            $descendantIds,
-            $this->systemStore->getSystem(),
-            null
-        );
-
-        $unauthorizedDescendantIds = \array_diff($descendantIds, $authorizedDescendantIds);
-        if (!empty($unauthorizedDescendantIds)) {
-            throw new InsufficientDescendantPermissionsException(
-                \count($unauthorizedDescendantIds),
-                PermissionTypes::DELETE
-            );
-        }
     }
 
     private function checkDescendantsForDependencyWarning(PageInterface $page): void
@@ -178,17 +132,5 @@ final class RemovePageMessageHandler
         \krsort($grouped);
 
         return \array_values($grouped);
-    }
-
-    private function getCurrentUser(): ?UserInterface
-    {
-        if (null === $this->security) {
-            return null;
-        }
-
-        /** @var UserInterface|null $user */
-        $user = $this->security->getUser();
-
-        return $user;
     }
 }
